Move plugins directory
