+++ components/DeadlinePicker.tsx
@@ -7,7 +7,7 @@ import {
   closeDeadlineCalendar,
   setDeadline,
 } from '@/store/todoSlice';
-import dayjs from '@/plugins/dayjs';
+import dayjs from '@/lib/plugins/dayjs';
 import styles from './ListOption.module.scss'; // shared
 import DeadlineCalendar from '@/components/DeadlineCalendar';
